Adding new test case. Cleaning up names and adding comments to test files to specify the test case that the class covers
