+++ src/test/java/test/thread/parallelization/ParallelByMethodsTestCase2Scenario1.java
@@ -5,12 +5,6 @@ import org.testng.TestNG;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import org.testng.xml.XmlSuite;
-import org.testng.xml.XmlTest;
-
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 
 import test.thread.parallelization.TestNgRunStateTracker.EventLog;
 
@@ -20,59 +14,67 @@ import test.thread.parallelization.sample.TestClassCSixMethodsWithNoDepsSample;
 import test.thread.parallelization.sample.TestClassDThreeMethodsWithNoDepsSample;
 import test.thread.parallelization.sample.TestClassEFiveMethodsWithNoDepsSample;
 import test.thread.parallelization.sample.TestClassFSixMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassGThreeMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassHFourMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassIFiveMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassJFourMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassKFiveMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassLThreeMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassMFourMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassNFiveMethodsWithNoDepsSample;
-import test.thread.parallelization.sample.TestClassOSixMethodsWithNoDepsSample;
+
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 import static org.testng.Assert.assertEquals;
 
-import static test.thread.parallelization.TestNgRunStateTracker.getAllEventLogsForSuite;
-
 import static test.thread.parallelization.TestNgRunStateTracker.getAllSuiteLevelEventLogs;
+import static test.thread.parallelization.TestNgRunStateTracker.getAllSuiteListenerStartEventLogs;
 import static test.thread.parallelization.TestNgRunStateTracker.getAllTestLevelEventLogs;
 import static test.thread.parallelization.TestNgRunStateTracker.getAllTestMethodLevelEventLogs;
 import static test.thread.parallelization.TestNgRunStateTracker.getSuiteAndTestLevelEventLogsForSuite;
 import static test.thread.parallelization.TestNgRunStateTracker.getSuiteLevelEventLogsForSuite;
 import static test.thread.parallelization.TestNgRunStateTracker.getSuiteListenerFinishEventLog;
 import static test.thread.parallelization.TestNgRunStateTracker.getSuiteListenerStartEventLog;
-import static test.thread.parallelization.TestNgRunStateTracker.getAllSuiteListenerStartEventLogs;
-
 import static test.thread.parallelization.TestNgRunStateTracker.getTestLevelEventLogsForSuite;
 import static test.thread.parallelization.TestNgRunStateTracker.getTestLevelEventLogsForTest;
 import static test.thread.parallelization.TestNgRunStateTracker.getTestListenerFinishEventLog;
-import static test.thread.parallelization.TestNgRunStateTracker.getTestListenerStartEventLog;
 
+import static test.thread.parallelization.TestNgRunStateTracker.getTestListenerStartEventLog;
 import static test.thread.parallelization.TestNgRunStateTracker.getTestMethodLevelEventLogsForSuite;
 import static test.thread.parallelization.TestNgRunStateTracker.getTestMethodLevelEventLogsForTest;
-
 import static test.thread.parallelization.TestNgRunStateTracker.reset;
 
-//Verify complex test run with multiple suites, tests and test classes and a thread count which is less than the number
-//of test methods to be executed.
-public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParallelizationTest {
+/**
+ * This class covers PTP_TC_2, Scenario 1 in the Parallelization Test Plan.
+ *
+ * Test Case Summary: Parallel by methods mode with parallel test suites, no dependencies and no factories or data
+ *                    providers.
+ *
+ * Scenario Description: Two suites with 1 and 2 tests respectively. One test for a suite shall consist of a single
+ *                       test class while the rest shall consist of more than one test class.
+ *
+ * 1) The suite thread pool is 2, so neither suite should have to wait for the other to complete execution
+ * 2) For both suites, the thread count and parallel mode are specified at the suite level
+ * 3) The thread count is less than the number of test methods in a test for one of the suites, so some methods will
+ *    have to wait the active thread count to drop below the maximum thread count before they can begin execution.
+ * 4) The thread count is more than the number of test methods for the rest of the tests, ensuring that none of their
+ *    test methods should have to wait for any other method to complete execution
+ * 5) There are NO configuration methods
+ * 6) All test methods pass
+ * 7) NO ordering is specified
+ * 8) group-by-instances is NOT set
+ * 9) There are no method exclusions
+ */
+public class ParallelByMethodsTestCase2Scenario1 extends BaseParallelizationTest {
+
     private static final String SUITE_A = "TestSuiteA";
     private static final String SUITE_B = "TestSuiteB";
-    private static final String SUITE_C = "TestSuiteC";
-    
+
     private static final String SUITE_A_TEST_A = "TestSuiteA-TwoTestClassTest";
 
     private static final String SUITE_B_TEST_A = "TestSuiteB-SingleTestClassTest";
     private static final String SUITE_B_TEST_B = "TestSuiteB-ThreeTestClassTest";
-    
-    private static final String SUITE_C_TEST_A = "TestSuiteC-ThreeTestClassTest";
-    private static final String SUITE_C_TEST_B = "TestSuiteC-TwoTestClassTest";
-    private static final String SUITE_C_TEST_C = "TestSuiteC-FourTestClassTest";
+
+    private static final int THREAD_POOL_SIZE = 2;
 
     private Map<String, Long> expectedSuiteExecutionTimes = new HashMap<>();
     private Map<String, Long> expectedTestExecutionTimes = new HashMap<>();
 
-    private Map<String, List<EventLog>> suiteEventLogsMap = new HashMap<>();
     private Map<String, List<EventLog>> testEventLogsMap = new HashMap<>();
 
     private List<EventLog> suiteLevelEventLogs;
@@ -89,28 +91,16 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
     private List<EventLog> suiteTwoTestLevelEventLogs;
     private List<EventLog> suiteTwoTestMethodLevelEventLogs;
 
-    private List<EventLog> suiteThreeSuiteAndTestLevelEventLogs;
-    private List<EventLog> suiteThreeSuiteLevelEventLogs;
-    private List<EventLog> suiteThreeTestLevelEventLogs;
-    private List<EventLog> suiteThreeTestMethodLevelEventLogs;
-
     private List<EventLog> suiteOneTestOneTestMethodLevelEventLogs;
 
     private List<EventLog> suiteTwoTestOneTestMethodLevelEventLogs;
     private List<EventLog> suiteTwoTestTwoTestMethodLevelEventLogs;
-
-    private List<EventLog> suiteThreeTestOneTestMethodLevelEventLogs;
-    private List<EventLog> suiteThreeTestTwoTestMethodLevelEventLogs;
-    private List<EventLog> suiteThreeTestThreeTestMethodLevelEventLogs;
 
     private EventLog suiteOneSuiteListenerOnStartEventLog;
     private EventLog suiteOneSuiteListenerOnFinishEventLog;
 
     private EventLog suiteTwoSuiteListenerOnStartEventLog;
     private EventLog suiteTwoSuiteListenerOnFinishEventLog;
-
-    private EventLog suiteThreeSuiteListenerOnStartEventLog;
-    private EventLog suiteThreeSuiteListenerOnFinishEventLog;
 
     private EventLog suiteOneTestOneListenerOnStartEventLog;
     private EventLog suiteOneTestOneListenerOnFinishEventLog;
@@ -120,23 +110,12 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
     private EventLog suiteTwoTestTwoListenerOnStartEventLog;
     private EventLog suiteTwoTestTwoListenerOnFinishEventLog;
 
-    private EventLog suiteThreeTestOneListenerOnStartEventLog;
-    private EventLog suiteThreeTestOneListenerOnFinishEventLog;
-    private EventLog suiteThreeTestTwoListenerOnStartEventLog;
-    private EventLog suiteThreeTestTwoListenerOnFinishEventLog;
-    private EventLog suiteThreeTestThreeListenerOnStartEventLog;
-    private EventLog suiteThreeTestThreeListenerOnFinishEventLog;
-
     @BeforeClass
-    public void complexMultipleSequentialSuites() {
+    public void setUp() {
         reset();
 
         XmlSuite suiteOne = createXmlSuite(SUITE_A);
         XmlSuite suiteTwo = createXmlSuite(SUITE_B);
-        XmlSuite suiteThree = createXmlSuite(SUITE_C);
-
-        suiteOne.setParallel(XmlSuite.ParallelMode.METHODS);
-        suiteOne.setThreadCount(3);
 
         createXmlTest(suiteOne, SUITE_A_TEST_A, TestClassAFiveMethodsWithNoDepsSample.class,
                 TestClassCSixMethodsWithNoDepsSample.class);
@@ -144,63 +123,28 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
         createXmlTest(suiteTwo, SUITE_B_TEST_B, TestClassDThreeMethodsWithNoDepsSample.class,
                 TestClassBFourMethodsWithNoDepsSample.class, TestClassFSixMethodsWithNoDepsSample.class);
 
+        suiteOne.setParallel(XmlSuite.ParallelMode.METHODS);
+        suiteOne.setThreadCount(3);
         suiteTwo.setParallel(XmlSuite.ParallelMode.METHODS);
+        suiteTwo.setThreadCount(14);
 
-        for(XmlTest test : suiteTwo.getTests()) {
-            if(test.getName().equals(SUITE_B_TEST_A)) {
-                test.setThreadCount(6);
-            } else {
-                test.setThreadCount(20);
-            }
-        }
-
-        createXmlTest(suiteThree, SUITE_C_TEST_A, TestClassGThreeMethodsWithNoDepsSample.class,
-                TestClassHFourMethodsWithNoDepsSample.class, TestClassIFiveMethodsWithNoDepsSample.class);
-        createXmlTest(suiteThree, SUITE_C_TEST_B, TestClassJFourMethodsWithNoDepsSample.class,
-                TestClassKFiveMethodsWithNoDepsSample.class);
-        createXmlTest(suiteThree, SUITE_C_TEST_C, TestClassLThreeMethodsWithNoDepsSample.class,
-                TestClassMFourMethodsWithNoDepsSample.class, TestClassNFiveMethodsWithNoDepsSample.class,
-                TestClassOSixMethodsWithNoDepsSample.class);
-
-        for(XmlTest test : suiteThree.getTests()) {
-            test.setParallel(XmlSuite.ParallelMode.METHODS);
-
-            switch(test.getName()) {
-                case SUITE_C_TEST_A:
-                    test.setThreadCount(10);
-                    break;
-                case SUITE_C_TEST_B:
-                    test.setThreadCount(5);
-                    break;
-                default:
-                    test.setThreadCount(12);
-                    break;
-            }
-        }
 
         addParams(suiteOne, SUITE_A, SUITE_A_TEST_A, "1");
 
         addParams(suiteTwo, SUITE_B, SUITE_B_TEST_A, "1");
         addParams(suiteTwo, SUITE_B, SUITE_B_TEST_B, "1");
 
-        addParams(suiteThree, SUITE_C, SUITE_C_TEST_A, "1");
-        addParams(suiteThree, SUITE_C, SUITE_C_TEST_B, "1");
-        addParams(suiteThree, SUITE_C, SUITE_C_TEST_C, "1");
-
-        TestNG tng = create(suiteOne, suiteTwo, suiteThree);
+        TestNG tng = create(suiteOne, suiteTwo);
+        tng.setSuiteThreadPoolSize(2);
         tng.addListener((ITestNGListener) new TestNgRunStateListener());
 
         tng.run();
 
         expectedSuiteExecutionTimes.put(SUITE_A, 10_000L);
         expectedSuiteExecutionTimes.put(SUITE_B, 7_000L);
-        expectedSuiteExecutionTimes.put(SUITE_C, 16_000L);
 
         expectedTestExecutionTimes.put(SUITE_B_TEST_A, 3_000L);
         expectedTestExecutionTimes.put(SUITE_B_TEST_B, 3_000L);
-        expectedTestExecutionTimes.put(SUITE_C_TEST_A, 5_000L);
-        expectedTestExecutionTimes.put(SUITE_C_TEST_B, 5_000L);
-        expectedTestExecutionTimes.put(SUITE_C_TEST_C, 5_000L);
 
         suiteLevelEventLogs = getAllSuiteLevelEventLogs();
         testLevelEventLogs = getAllTestLevelEventLogs();
@@ -214,42 +158,22 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
         suiteTwoSuiteLevelEventLogs = getSuiteLevelEventLogsForSuite(SUITE_B);
         suiteTwoTestLevelEventLogs = getTestLevelEventLogsForSuite(SUITE_B);
 
-        suiteThreeSuiteAndTestLevelEventLogs = getSuiteAndTestLevelEventLogsForSuite(SUITE_C);
-        suiteThreeSuiteLevelEventLogs = getSuiteLevelEventLogsForSuite(SUITE_C);
-        suiteThreeTestLevelEventLogs = getTestLevelEventLogsForSuite(SUITE_C);
-
-        suiteEventLogsMap.put(SUITE_A, getAllEventLogsForSuite(SUITE_A));
-        suiteEventLogsMap.put(SUITE_B, getAllEventLogsForSuite(SUITE_B));
-        suiteEventLogsMap.put(SUITE_C, getAllEventLogsForSuite(SUITE_C));
-
         suiteOneTestMethodLevelEventLogs = getTestMethodLevelEventLogsForSuite(SUITE_A);
         suiteTwoTestMethodLevelEventLogs = getTestMethodLevelEventLogsForSuite(SUITE_B);
-        suiteThreeTestMethodLevelEventLogs = getTestMethodLevelEventLogsForSuite(SUITE_C);
 
         suiteOneTestOneTestMethodLevelEventLogs = getTestMethodLevelEventLogsForTest(SUITE_A, SUITE_A_TEST_A);
 
         suiteTwoTestOneTestMethodLevelEventLogs = getTestMethodLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_A);
         suiteTwoTestTwoTestMethodLevelEventLogs = getTestMethodLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_B);
 
-        suiteThreeTestOneTestMethodLevelEventLogs = getTestMethodLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_A);
-        suiteThreeTestTwoTestMethodLevelEventLogs = getTestMethodLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_B);
-        suiteThreeTestThreeTestMethodLevelEventLogs = getTestMethodLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_C);
-
         testEventLogsMap.put(SUITE_B_TEST_A, getTestLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_A));
         testEventLogsMap.put(SUITE_B_TEST_B, getTestLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_B));
-
-        testEventLogsMap.put(SUITE_C_TEST_A, getTestLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_A));
-        testEventLogsMap.put(SUITE_C_TEST_B, getTestLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_B));
-        testEventLogsMap.put(SUITE_C_TEST_C, getTestLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_C));
 
         suiteOneSuiteListenerOnStartEventLog = getSuiteListenerStartEventLog(SUITE_A);
         suiteOneSuiteListenerOnFinishEventLog = getSuiteListenerFinishEventLog(SUITE_A);
 
         suiteTwoSuiteListenerOnStartEventLog = getSuiteListenerStartEventLog(SUITE_B);
         suiteTwoSuiteListenerOnFinishEventLog = getSuiteListenerFinishEventLog(SUITE_B);
-
-        suiteThreeSuiteListenerOnStartEventLog = getSuiteListenerStartEventLog(SUITE_C);
-        suiteThreeSuiteListenerOnFinishEventLog = getSuiteListenerFinishEventLog(SUITE_C);
 
         suiteOneTestOneListenerOnStartEventLog = getTestListenerStartEventLog(SUITE_A, SUITE_A_TEST_A);
         suiteOneTestOneListenerOnFinishEventLog = getTestListenerFinishEventLog(SUITE_A, SUITE_A_TEST_A);
@@ -260,25 +184,19 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
         suiteTwoTestTwoListenerOnStartEventLog = getTestListenerStartEventLog(SUITE_B, SUITE_B_TEST_B);
         suiteTwoTestTwoListenerOnFinishEventLog = getTestListenerFinishEventLog(SUITE_B, SUITE_B_TEST_B);
 
-        suiteThreeTestOneListenerOnStartEventLog = getTestListenerStartEventLog(SUITE_C, SUITE_C_TEST_A);
-        suiteThreeTestOneListenerOnFinishEventLog = getTestListenerFinishEventLog(SUITE_C, SUITE_C_TEST_A);
-        suiteThreeTestTwoListenerOnStartEventLog = getTestListenerStartEventLog(SUITE_C, SUITE_C_TEST_B);
-        suiteThreeTestTwoListenerOnFinishEventLog = getTestListenerFinishEventLog(SUITE_C, SUITE_C_TEST_B);
-        suiteThreeTestThreeListenerOnStartEventLog = getTestListenerStartEventLog(SUITE_C, SUITE_C_TEST_C);
-        suiteThreeTestThreeListenerOnFinishEventLog = getTestListenerFinishEventLog(SUITE_C, SUITE_C_TEST_C);
     }
 
     //Verifies that the expected number of suite, test and test method level events were logged for each of the three
     //suites.
     @Test
     public void sanityCheck() {
-        assertEquals(suiteLevelEventLogs.size(), 6, "There should be 6 suite level events logged for " + SUITE_A +
-                ", " + SUITE_B + " and " + SUITE_C + ": " + suiteLevelEventLogs);
-        assertEquals(testLevelEventLogs.size(), 12, "There should be 12 test level events logged for " + SUITE_A +
-                ", " + SUITE_B + " and " + SUITE_C + ": " + testLevelEventLogs);
+        assertEquals(suiteLevelEventLogs.size(), 4, "There should be 4 suite level events logged for " + SUITE_A +
+                ", " + SUITE_B + ": " + suiteLevelEventLogs);
+        assertEquals(testLevelEventLogs.size(), 6, "There should be 6 test level events logged for " + SUITE_A +
+                ", " + SUITE_B + ": " + testLevelEventLogs);
 
-        assertEquals(testMethodLevelEventLogs.size(), 204, "There should 204 test method level events logged for " +
-                SUITE_A + ", " + SUITE_B + " and " + SUITE_C + ": " + testMethodLevelEventLogs);
+        assertEquals(testMethodLevelEventLogs.size(), 87, "There should 87 test method level events logged for " +
+                SUITE_A + ", " + SUITE_B + ": " + testMethodLevelEventLogs);
 
         assertEquals(suiteOneSuiteLevelEventLogs.size(), 2, "There should be 2 suite level events logged for " +
                 SUITE_A + ": " + suiteOneSuiteLevelEventLogs);
@@ -293,43 +211,25 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
                 ": " + suiteTwoTestLevelEventLogs);
         assertEquals(suiteTwoTestMethodLevelEventLogs.size(), 54, "There should be 54 test method level events " +
                 "logged for " + SUITE_B + ": " + suiteTwoTestMethodLevelEventLogs);
-
-        assertEquals(suiteThreeSuiteLevelEventLogs.size(), 2, "There should be 2 suite level events logged for " +
-                SUITE_C + ": " + suiteThreeSuiteLevelEventLogs);
-        assertEquals(suiteThreeTestLevelEventLogs.size(), 6, "There should be 6 test level events logged for " +
-                SUITE_C + ": " + suiteThreeTestLevelEventLogs);
-        assertEquals(suiteThreeTestMethodLevelEventLogs.size(), 117, "There should be 87 test method level events " +
-                "logged for " + SUITE_C + ": " + suiteThreeTestMethodLevelEventLogs);
-
     }
 
-    //Verify that all the events in the second suite and third suites run have timestamps later than the suite
-    //listener's onFinish event for the first suite run.
-    //Verify that all the events in the third suite run have timestamps later than the suite listener's onFinish
-    //event for the second suite run.
-    //Verify that all suite level events run in the same thread
+    //Verify that the suites run in parallel by checking that the suite and test level events for both suites have
+    //overlapping timestamps. Verify that there are two separate threads executing the suite-level and test-level
+    //events for each suite.
     @Test
-    public void verifySuitesRunSequentiallyInSameThread() {
-        verifySequentialSuites(suiteLevelEventLogs, suiteEventLogsMap);
+    public void verifyThatSuitesRunInParallelThreads() {
+        verifyParallelSuitesWithUnequalExecutionTimes(suiteLevelEventLogs, THREAD_POOL_SIZE);
     }
 
-    //For all suites, verify that the test level events run sequentially because the parallel mode is by methods only.
     @Test
-    public void verifySuiteAndTestLevelEventsRunInSequentialOrderForIndividualSuites() {
-
+    public void verifyTestLevelEventsRunInSequentialOrderForIndividualSuites() {
         verifySequentialTests(suiteOneSuiteAndTestLevelEventLogs, suiteOneTestLevelEventLogs,
                 suiteOneSuiteListenerOnStartEventLog, suiteOneSuiteListenerOnFinishEventLog);
 
         verifySequentialTests(suiteTwoSuiteAndTestLevelEventLogs, suiteTwoTestLevelEventLogs,
                 suiteTwoSuiteListenerOnStartEventLog, suiteTwoSuiteListenerOnFinishEventLog);
-
-        verifySequentialTests(suiteThreeSuiteAndTestLevelEventLogs, suiteThreeTestLevelEventLogs,
-                suiteThreeSuiteListenerOnStartEventLog, suiteThreeSuiteListenerOnFinishEventLog);
     }
 
-    //Verify that there is only a single test class instance associated with each of the test methods from the sample
-    //classes for every test in all the suites.
-    //Verify that the same test class instance is associated with each of the test methods from the sample test class
     @Test
     public void verifyOnlyOneInstanceOfTestClassForAllTestMethodsForAllSuites() {
 
@@ -377,62 +277,6 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
                         TestClassFSixMethodsWithNoDepsSample.class
                 )
         );
-
-        verifyNumberOfInstancesOfTestClassesForMethods(
-                SUITE_C,
-                SUITE_C_TEST_A,
-                Arrays.asList(
-                        TestClassGThreeMethodsWithNoDepsSample.class,
-                        TestClassHFourMethodsWithNoDepsSample.class,
-                        TestClassIFiveMethodsWithNoDepsSample.class
-                ),
-                1
-        );
-
-        verifySameInstancesOfTestClassesAssociatedWithMethods(
-                SUITE_C,
-                SUITE_C_TEST_A,
-                Arrays.asList(
-                        TestClassGThreeMethodsWithNoDepsSample.class,
-                        TestClassHFourMethodsWithNoDepsSample.class,
-                        TestClassIFiveMethodsWithNoDepsSample.class
-                )
-        );
-
-        verifyNumberOfInstancesOfTestClassesForMethods(
-                SUITE_C,
-                SUITE_C_TEST_B,
-                Arrays.asList(TestClassJFourMethodsWithNoDepsSample.class, TestClassKFiveMethodsWithNoDepsSample.class),
-                1);
-
-        verifySameInstancesOfTestClassesAssociatedWithMethods(
-                SUITE_C,
-                SUITE_C_TEST_B,
-                Arrays.asList(TestClassJFourMethodsWithNoDepsSample.class, TestClassKFiveMethodsWithNoDepsSample.class)
-        );
-
-        verifyNumberOfInstancesOfTestClassesForMethods(
-                SUITE_C,
-                SUITE_C_TEST_C,
-                Arrays.asList(
-                        TestClassLThreeMethodsWithNoDepsSample.class,
-                        TestClassMFourMethodsWithNoDepsSample.class,
-                        TestClassNFiveMethodsWithNoDepsSample.class,
-                        TestClassOSixMethodsWithNoDepsSample.class
-                ),
-                1
-        );
-
-        verifySameInstancesOfTestClassesAssociatedWithMethods(
-                SUITE_C,
-                SUITE_C_TEST_C,
-                Arrays.asList(
-                        TestClassLThreeMethodsWithNoDepsSample.class,
-                        TestClassMFourMethodsWithNoDepsSample.class,
-                        TestClassNFiveMethodsWithNoDepsSample.class,
-                        TestClassOSixMethodsWithNoDepsSample.class
-                )
-        );
     }
 
     //Verify that the test method listener's onTestStart method runs after the test listener's onStart method for
@@ -462,31 +306,6 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
                         suiteTwoTestTwoListenerOnStartEventLog + ". Test listener onFinish event log: " +
                         suiteTwoTestTwoListenerOnFinishEventLog + ". Test method level event logs: " +
                         suiteTwoTestTwoTestMethodLevelEventLogs);
-
-        verifyEventsOccurBetween(suiteThreeTestOneListenerOnStartEventLog, suiteThreeTestOneTestMethodLevelEventLogs,
-                suiteThreeTestOneListenerOnFinishEventLog,  "All of the test method level event logs for " +
-                        SUITE_C_TEST_A + " should have timestamps between the test listener's onStart and onFinish " +
-                        "event logs for " + SUITE_C_TEST_A + ". Test listener onStart event log: " +
-                        suiteThreeTestOneListenerOnStartEventLog + ". Test listener onFinish event log: " +
-                        suiteThreeTestOneListenerOnFinishEventLog + ". Test method level event logs: " +
-                        suiteThreeTestOneTestMethodLevelEventLogs);
-
-        verifyEventsOccurBetween(suiteThreeTestTwoListenerOnStartEventLog, suiteThreeTestTwoTestMethodLevelEventLogs,
-                suiteThreeTestTwoListenerOnFinishEventLog,  "All of the test method level event logs for " +
-                        SUITE_C_TEST_B + " should have timestamps between the test listener's onStart and onFinish " +
-                        "event logs for " + SUITE_C_TEST_B + ". Test listener onStart event log: " +
-                        suiteThreeTestTwoListenerOnStartEventLog + ". Test listener onFinish event log: " +
-                        suiteThreeTestTwoListenerOnFinishEventLog + ". Test method level event logs: " +
-                        suiteThreeTestTwoTestMethodLevelEventLogs);
-
-        verifyEventsOccurBetween(suiteThreeTestThreeListenerOnStartEventLog, suiteThreeTestThreeTestMethodLevelEventLogs,
-                suiteThreeTestThreeListenerOnFinishEventLog,  "All of the test method level event logs for " +
-                        SUITE_C_TEST_C + " should have timestamps between the test listener's onStart and onFinish " +
-                        "event logs for " + SUITE_C_TEST_C + ". Test listener onStart event log: " +
-                        suiteThreeTestThreeListenerOnStartEventLog + ". Test listener onFinish event log: " +
-                        suiteThreeTestThreeListenerOnFinishEventLog + ". Test method level event logs: " +
-                        suiteThreeTestThreeTestMethodLevelEventLogs);
-
     }
 
     //Verifies that the method level events all run in different threads from the test and suite level events.
@@ -505,6 +324,8 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
 
         verifyEventsForTestMethodsRunInTheSameThread(TestClassAFiveMethodsWithNoDepsSample.class, SUITE_A,
                 SUITE_A_TEST_A);
+        verifyEventsForTestMethodsRunInTheSameThread(TestClassAFiveMethodsWithNoDepsSample.class, SUITE_A,
+                SUITE_A_TEST_A);
         verifyEventsForTestMethodsRunInTheSameThread(TestClassCSixMethodsWithNoDepsSample.class, SUITE_A,
                 SUITE_A_TEST_A);
 
@@ -517,37 +338,17 @@ public class ParallelByMethodsMultipleSuitesTestsClassesTest extends BaseParalle
                 SUITE_B_TEST_B);
         verifyEventsForTestMethodsRunInTheSameThread(TestClassFSixMethodsWithNoDepsSample.class, SUITE_B,
                 SUITE_B_TEST_B);
-
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassGThreeMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_A);
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassHFourMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_A);
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassIFiveMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_A);
-
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassJFourMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_B);
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassKFiveMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_B);
-
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassLThreeMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_C);
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassMFourMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_C);
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassNFiveMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_C);
-        verifyEventsForTestMethodsRunInTheSameThread(TestClassOSixMethodsWithNoDepsSample.class, SUITE_C,
-                SUITE_C_TEST_C);
     }
 
-    //Verify that the methods are run in separate threads.
+    //Verify that the methods are run in separate threads in true parallel fashion by checking that the start and run
+    //times of events that should be run simultaneously start basically at the same time using the timestamps and the
+    //known values of the wait time specified for the event. Verify that the thread IDs of parallel events are
+    //different.
     @Test
     public void verifyThatTestMethodsRunInParallelThreads() {
         verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_A, SUITE_A_TEST_A), SUITE_A_TEST_A, 3);
-        verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_A), SUITE_B_TEST_A, 6);
-        verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_B), SUITE_B_TEST_B, 20);
-        verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_A), SUITE_C_TEST_A, 10);
-        verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_B), SUITE_C_TEST_B, 5);
-        verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_C, SUITE_C_TEST_C), SUITE_C_TEST_C, 12);
+        verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_A), SUITE_B_TEST_A, 14);
+        verifySimultaneousTestMethods(getTestMethodLevelEventLogsForTest(SUITE_B, SUITE_B_TEST_B), SUITE_B_TEST_B, 14);
     }
+
 }
