+++ src/test/java/test/thread/parallelization/ParallelByMethodsTestCase3Scenario1.java
@@ -1,20 +1,22 @@
 package test.thread.parallelization;
 
 import com.google.common.collect.Multimap;
+
 import org.testng.ITestNGListener;
 import org.testng.TestNG;
+
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
+
 import org.testng.xml.XmlSuite;
 
 import test.thread.parallelization.TestNgRunStateTracker.EventLog;
-import test.thread.parallelization.sample.TestClassAFiveMethodsWithNoDepsSample;
+import test.thread.parallelization.sample.TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.testng.Assert.assertEquals;
-
 import static test.thread.parallelization.TestNgRunStateTracker.getAllSuiteAndTestLevelEventLogs;
 import static test.thread.parallelization.TestNgRunStateTracker.getAllSuiteLevelEventLogs;
 import static test.thread.parallelization.TestNgRunStateTracker.getAllTestLevelEventLogs;
@@ -25,13 +27,29 @@ import static test.thread.parallelization.TestNgRunStateTracker.getTestListenerF
 import static test.thread.parallelization.TestNgRunStateTracker.getTestListenerStartEventLog;
 import static test.thread.parallelization.TestNgRunStateTracker.getTestListenerStartThreadId;
 import static test.thread.parallelization.TestNgRunStateTracker.getTestMethodEventLogsForMethod;
-
 import static test.thread.parallelization.TestNgRunStateTracker.reset;
 
-//Verify simple test run with a single suite which consists of a single test with a single test class. The thread count
-//is sufficient to run all methods in parallel at once, so no methods should be queued.
-public class ParallelByMethodsSingleSuiteSingleTestSingleClassTest extends BaseParallelizationTest {
-
+/**
+ * This class covers PTP_TC_3, Scenario 1 in the Parallelization Test Plan.
+ *
+ * Test Case Summary: Parallel by methods mode with sequential test suites using a non-parallel data provider but no
+ *                    dependencies and no factories.
+ *
+ * Scenario Description: Single suite with a single test consisting of a single test class with five methods with a data
+ *                       provider specifying 3 sets of data
+ *
+ * 1) Thread count and parallel mode are specified at the suite level
+ * 2) The thread count is equal to the number of test methods times 3, the number of times each method will be invoked
+ *    with a data set from the data provider. Expectation is that only 5 threads will be spawned, one for each of the
+ *    methods and that in this thread, the method will be invoked 3 times, once for each set of data from the data
+ *    provider.
+ * 3) There are NO configuration methods
+ * 4) All test methods pass
+ * 5) NO ordering is specified
+ * 6) group-by-instances is NOT set
+ * 7) There are no method exclusions
+ */
+public class ParallelByMethodsTestCase3Scenario1 extends BaseParallelizationTest {
     private static final String SUITE = "SingleTestSuite";
     private static final String TEST = "SingleTestClassTest";
 
@@ -55,18 +73,19 @@ public class ParallelByMethodsSingleSuiteSingleTestSingleClassTest extends BaseP
     private Multimap<Object, EventLog> testMethodEEventLogs;
 
     @BeforeClass
-    public void singleSuiteSingleTestSingleTestClass() {
+    public void setUp() {
         reset();
 
         XmlSuite suite = createXmlSuite(SUITE);
         suite.setParallel(XmlSuite.ParallelMode.METHODS);
-        suite.setThreadCount(5);
+        suite.setThreadCount(15);
 
-        createXmlTest(suite, TEST, TestClassAFiveMethodsWithNoDepsSample.class);
+        createXmlTest(suite, TEST, TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class);
 
-        addParams(suite, SUITE, TEST, "1");
+        addParams(suite, SUITE, TEST, "1", "paramOne,paramTwo,paramThree");
 
         TestNG tng = create(suite);
+
         tng.addListener((ITestNGListener)new TestNgRunStateListener());
 
         tng.run();
@@ -77,15 +96,15 @@ public class ParallelByMethodsSingleSuiteSingleTestSingleClassTest extends BaseP
         testMethodLevelEventLogs = getAllTestMethodLevelEventLogs();
 
         testMethodAEventLogs = getTestMethodEventLogsForMethod(SUITE, TEST,
-                TestClassAFiveMethodsWithNoDepsSample.class.getCanonicalName(), "testMethodA");
+                TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class.getCanonicalName(), "testMethodA");
         testMethodBEventLogs = getTestMethodEventLogsForMethod(SUITE, TEST,
-                TestClassAFiveMethodsWithNoDepsSample.class.getCanonicalName(), "testMethodB");
+                TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class.getCanonicalName(), "testMethodB");
         testMethodCEventLogs = getTestMethodEventLogsForMethod(SUITE, TEST,
-                TestClassAFiveMethodsWithNoDepsSample.class.getCanonicalName(), "testMethodC");
+                TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class.getCanonicalName(), "testMethodC");
         testMethodDEventLogs = getTestMethodEventLogsForMethod(SUITE, TEST,
-                TestClassAFiveMethodsWithNoDepsSample.class.getCanonicalName(), "testMethodD");
+                TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class.getCanonicalName(), "testMethodD");
         testMethodEEventLogs = getTestMethodEventLogsForMethod(SUITE, TEST,
-                TestClassAFiveMethodsWithNoDepsSample.class.getCanonicalName(), "testMethodE");
+                TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class.getCanonicalName(), "testMethodE");
 
         suiteListenerOnStartEventLog = getSuiteListenerStartEventLog(SUITE);
         suiteListenerOnFinishEventLog = getSuiteListenerFinishEventLog(SUITE);
@@ -103,7 +122,7 @@ public class ParallelByMethodsSingleSuiteSingleTestSingleClassTest extends BaseP
                 suiteLevelEventLogs);
         assertEquals(testLevelEventLogs.size(), 2, "There should be 2 test level events logged for " + SUITE + ": " +
                 testLevelEventLogs);
-        assertEquals(testMethodLevelEventLogs.size(), 15, "There should be 15 test method level event logged for " +
+        assertEquals(testMethodLevelEventLogs.size(), 45, "There should be 15 test method level event logged for " +
                 SUITE + ": " + testMethodLevelEventLogs);
     }
 
@@ -134,13 +153,13 @@ public class ParallelByMethodsSingleSuiteSingleTestSingleClassTest extends BaseP
         verifyNumberOfInstancesOfTestClassForMethods(
                 SUITE,
                 TEST,
-                TestClassAFiveMethodsWithNoDepsSample.class,
+                TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class,
                 1);
 
         verifySameInstancesOfTestClassAssociatedWithMethods(
                 SUITE,
                 TEST,
-                TestClassAFiveMethodsWithNoDepsSample.class);
+                TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.class);
     }
 
     //Verifies that all the test method level events execute between the test listener onStart and onFinish methods
@@ -204,4 +223,5 @@ public class ParallelByMethodsSingleSuiteSingleTestSingleClassTest extends BaseP
                         testMethodEEventLogs.get(testMethodEEventLogs.keySet().toArray()[0])
         );
     }
+
 }
