+++ src/test/java/test/thread/parallelization/sample/TestClassFiveMethodsWithDataProviderOnAllMethodsAndNoDepsSample.java
@@ -5,7 +5,6 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import test.thread.parallelization.TestNgRunStateTracker;
 
-import java.util.Arrays;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
